# src/main/resources/api-specs/product-api-specs.yml
openapi: 3.1.0
info:
  title: Cosmo Cats Intergalactic Marketplace API
  version: 1.0.0

servers:
  - url: http://localhost:8080

tags:
  - name: Products
    description: Product Operations

paths:

  /api/v1/products:
    get:
      tags: [Products]
      summary: Get All Products
      operationId: listProducts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseProductDto'

    post:
      tags: [Products]
      summary: Create New Product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestProductDto'
      responses:
        '201':
          description: Created Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseProductDto'
        '400':
          $ref: '#/components/responses/ValidationProblem'

  /api/v1/products/{id}:
    get:
      tags: [Products]
      summary: Get Product by ID
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseProductDto'
        '404':
          $ref: '#/components/responses/NotFoundProblem'

    put:
      tags: [Products]
      summary: Update Product
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestProductDto'
      responses:
        '200':
          description: Updated Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseProductDto'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'

    delete:
      tags: [Products]
      summary: Delete Product
      operationId: deleteProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Product Deleted
        '404':
          $ref: '#/components/responses/NotFoundProblem'

components:

  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Unique Product ID

  schemas:

    RequestProductDto:
      type: object
      required: [name, description, price]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        description:
          type: string
          minLength: 2
          maxLength: 500
        price:
          type: number
          minimum: 0
          maximum: 1000

    ResponseProductDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number

    Problem:
      description: RFC 9457 formatted error
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string

  responses:

    ValidationProblem:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'

    NotFoundProblem:
      description: Entity not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'